Move utils to their own folder
